board reboot and force start methods



See merge request !25

+++ Utils/Data.h
@@ -102,7 +102,7 @@ namespace Ph2_HwInterface
         }
         const Event* GetEvent ( const BeBoard* pBoard, int i ) const
         {
-            return ( ( i >= fEventList.size() ) ? nullptr : fEventList.at ( i ) );
+            return ( ( i >= (int)fEventList.size() ) ? nullptr : fEventList.at ( i ) );
         }
         const std::vector<Event*>& GetEvents ( const BeBoard* pBoard ) const
         {
